Lint sawtoothXO
Lints sawtoothXO app using ktlint. Renames the project from sawtooth_xo to sawtoothXO to comply with project naming conventions.
