+++ examples/xo_android_client/app/src/androidTest/java/bitwiseio/sawtooth/xo/ExampleInstrumentedTest.kt
@@ -1,0 +1,24 @@
+package bitwiseio.sawtooth.xo
+
+import android.support.test.InstrumentationRegistry
+import android.support.test.runner.AndroidJUnit4
+
+import org.junit.Test
+import org.junit.runner.RunWith
+
+import org.junit.Assert.assertEquals
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+@RunWith(AndroidJUnit4::class)
+class ExampleInstrumentedTest {
+    @Test
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getTargetContext()
+        assertEquals("io.bitwise.sawtooth_xo", appContext.packageName)
+    }
+}
