--- examples/xo_android_client/app/src/androidTest/java/io/bitwise/sawtooth_xo/ExampleInstrumentedTest.kt
@@ -1,24 +1,0 @@
-package io.bitwise.sawtooth_xo
-
-import android.support.test.InstrumentationRegistry
-import android.support.test.runner.AndroidJUnit4
-
-import org.junit.Test
-import org.junit.runner.RunWith
-
-import org.junit.Assert.*
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getTargetContext()
-        assertEquals("io.bitwise.sawtooth_xo", appContext.packageName)
-    }
-}
