+++ examples/xo_android_client/app/src/main/java/bitwiseio/sawtooth/xo/SettingsActivity.kt
@@ -1,0 +1,13 @@
+package bitwiseio.sawtooth.xo
+
+import android.os.Bundle
+import android.support.v7.app.AppCompatActivity
+
+class SettingsActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_settings)
+        setSupportActionBar(findViewById(R.id.action_menu_settings))
+    }
+}
