+++ examples/xo_android_client/app/src/main/java/bitwiseio/sawtooth/xo/SettingsFragment.kt
@@ -1,0 +1,10 @@
+package bitwiseio.sawtooth.xo
+
+import android.os.Bundle
+import android.support.v7.preference.PreferenceFragmentCompat
+
+class SettingsFragment : PreferenceFragmentCompat() {
+    override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
+        setPreferencesFromResource(R.xml.preferences, rootKey)
+    }
+}
