+++ examples/xo_android_client/app/src/main/java/bitwiseio/sawtooth/xo/models/Game.kt
@@ -1,0 +1,7 @@
+package bitwiseio.sawtooth.xo.models
+
+class Game(var name: String, var board: String, var gameState: String, var playerKey1: String, var playerKey2: String) {
+    override fun toString(): String {
+        return "Game(name=$name, board=$board, gameState=$gameState, playerKey1=$playerKey1, playerKey2=$playerKey2)"
+    }
+}
