+++ examples/xo_android_client/app/src/main/java/bitwiseio/sawtooth/xo/state/api/BatchListResponse.kt
@@ -1,0 +1,20 @@
+package bitwiseio.sawtooth.xo.state.api
+
+import com.google.gson.annotations.SerializedName
+
+/**
+ * Data class for interacting with the Sawtooth REST Api
+ */
+class BatchListResponse(
+    @field:SerializedName("link")
+    /** Returns the link to query the batch status
+     *
+     * @return String link
+     */
+    val link: String
+
+) {
+    override fun toString(): String {
+        return "BatchListResponse(link='$link')"
+    }
+}
