--- examples/xo_android_client/app/src/main/java/io/bitwise/sawtooth_xo/SettingsActivity.kt
@@ -1,13 +1,0 @@
-package io.bitwise.sawtooth_xo
-
-import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
-
-class SettingsActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_settings)
-        setSupportActionBar(findViewById(R.id.action_menu_settings))
-    }
-}
