--- examples/xo_android_client/app/src/main/java/io/bitwise/sawtooth_xo/SettingsFragment.kt
@@ -1,10 +1,0 @@
-package io.bitwise.sawtooth_xo
-
-import android.os.Bundle
-import android.support.v7.preference.PreferenceFragmentCompat
-
-class SettingsFragment : PreferenceFragmentCompat() {
-    override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
-        setPreferencesFromResource(R.xml.preferences, rootKey)
-    }
-}
