--- examples/xo_android_client/app/src/main/java/io/bitwise/sawtooth_xo/adapters/PagerAdapter.kt
@@ -1,29 +1,0 @@
-package io.bitwise.sawtooth_xo.adapters
-
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentManager
-import android.support.v4.app.FragmentPagerAdapter
-
-class PagerAdapter(fm: FragmentManager) : FragmentPagerAdapter(fm) {
-    private val mFragmentList = arrayListOf<Fragment>()
-    private val mFragmentTitleList = arrayListOf<String>()
-
-    override fun getItem(position: Int): Fragment {
-        return mFragmentList[position]
-    }
-
-    override fun getCount(): Int {
-        return mFragmentList.size
-    }
-
-    fun addFragment(fragment: Fragment, title: String) {
-        mFragmentList.add(fragment)
-        mFragmentTitleList.add(title)
-
-    }
-
-    override fun getPageTitle(position: Int): CharSequence? {
-        return mFragmentTitleList[position]
-    }
-
-}
