--- examples/xo_android_client/app/src/main/java/io/bitwise/sawtooth_xo/state/rest_api/BatchListResponse.kt
@@ -1,20 +1,0 @@
-package io.bitwise.sawtooth_xo.state.rest_api
-
-import com.google.gson.annotations.SerializedName
-
-/**
- * Data class for interacting with the Sawtooth REST Api
- */
-class BatchListResponse(
-    @field:SerializedName("link")
-    /** Returns the link to query the batch status
-     *
-     * @return String link
-     */
-    val link: String
-
-) {
-    override fun toString(): String {
-        return "BatchListResponse(link='$link')"
-    }
-}
