+++ examples/xo_android_client/app/src/main/res/layout/activity_settings.xml
@@ -20,7 +20,7 @@
     </android.support.design.widget.AppBarLayout>
     <fragment
             android:layout_width="0dp"
-            android:layout_height="wrap_content" android:name="io.bitwise.sawtooth_xo.SettingsFragment"
+            android:layout_height="wrap_content" android:name="bitwiseio.sawtooth.xo.SettingsFragment"
             android:id="@+id/fragment" android:layout_marginTop="8dp"
             app:layout_constraintTop_toBottomOf="@+id/appbar" app:layout_constraintEnd_toEndOf="parent"
             android:layout_marginEnd="8dp" android:layout_marginRight="8dp"
