+++ examples/xo_android_client/app/src/main/res/layout/fragment_game_list.xml
@@ -4,7 +4,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         xmlns:tools="http://schemas.android.com/tools"
         android:id="@+id/game_list"
-        android:name="io.bitwise.sawtooth_xo.GameListFragment"
+        android:name="bitwiseio.sawtooth.xo.GameListFragment"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:layoutManager="android.support.v7.widget.LinearLayoutManager"
