+++ examples/xo_android_client/app/src/test/java/bitwiseio/sawtooth/xo/ExampleUnitTest.kt
@@ -1,0 +1,17 @@
+package bitwiseio.sawtooth.xo
+
+import org.junit.Test
+
+import org.junit.Assert.assertEquals
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class ExampleUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        assertEquals(4, 2 + 2)
+    }
+}
